requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.3.1; python_version >= '3.5'
dj-database-url==0.5.0
django-heroku==0.3.1
django==3.1.4
gunicorn==20.0.4
psycopg2-binary==2.8.6
psycopg2==2.8.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2020.4
sqlparse==0.4.1; python_version >= '3.5'
whitenoise==5.2.0
djangorestframework==3.12.2